+++ src/cli.ts
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
 import yargs from "yargs";
+import { setAnalyticsDisabled } from "./analytics";
 import { LoggableError } from "./errors/LoggableError";
 import { InitializeMode, Initializer } from "./Initializer";
 import { Project } from "./Project";
@@ -73,8 +74,15 @@ const argv = yargs
 	// init
 	.option("init", {
 		choices: [InitializeMode.Game, InitializeMode.Model, InitializeMode.Plugin, InitializeMode.Package],
-		conflicts: ["w"],
+		conflicts: ["w", "no-analytics"],
 		type: "string",
+	})
+
+	// noAnalytics
+	.option("noAnalytics", {
+		conflicts: ["w", "init"],
+		description: "disables analytics globally",
+		type: "boolean",
 	})
 
 	// parse
@@ -82,7 +90,9 @@ const argv = yargs
 
 void (async () => {
 	try {
-		if (argv.init !== undefined) {
+		if (argv.noAnalytics !== undefined) {
+			await setAnalyticsDisabled(argv.noAnalytics);
+		} else if (argv.init !== undefined) {
 			await Initializer.init(argv.init as InitializeMode);
 		} else if (argv.watch === true) {
 			new Watcher(new Project(argv), argv.onSuccess).start();